+++ src/main/java/dmitriy/tsoy/russia/spendMoneyControlServerPart/controller/UserController.java
@@ -1,6 +1,7 @@
 package dmitriy.tsoy.russia.spendMoneyControlServerPart.controller;
 
 import dmitriy.tsoy.russia.spendMoneyControlServerPart.dto.UserDto;
+import dmitriy.tsoy.russia.spendMoneyControlServerPart.model.Record;
 import dmitriy.tsoy.russia.spendMoneyControlServerPart.model.User;
 import dmitriy.tsoy.russia.spendMoneyControlServerPart.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,18 +22,26 @@ public class UserController {
 
     @GetMapping()
     public ResponseEntity<List<User>> getAllUsers() {
-        return ResponseEntity.ok(userService.getAllUsers());
+        List<User> users = userService.getAllUsers();
+        return users.isEmpty()
+                ? new ResponseEntity<>(HttpStatus.NOT_FOUND)
+                : new ResponseEntity<>(users, HttpStatus.OK);
     }
 
     @GetMapping("{id}")
     public ResponseEntity<Optional<User>> getUserById(@PathVariable(value="id") long id) {
-        return ResponseEntity.ok(userService.getUserById(id));
+        Optional<User> user = userService.getUserById(id);
+        return user != null
+                ? new ResponseEntity<>(user, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PostMapping()
     public ResponseEntity<String> saveUser(@RequestBody User user) {
         userService.saveUser(user);
-        return new ResponseEntity<>("User successfully created", HttpStatus.CREATED);
+        return userService.getUserById(user.getId()) != null
+                ? new ResponseEntity<>("User saved successfully", HttpStatus.CREATED)
+                : new ResponseEntity<>(HttpStatus.SERVICE_UNAVAILABLE);
     }
 
     @PutMapping("{id}")
@@ -40,18 +49,32 @@ public class UserController {
                                              @RequestParam(value="username", required = false, defaultValue = "") String username,
                                              @RequestParam(value="age", required = false, defaultValue = "0") int age,
                                              @RequestParam(value="sex", required = false, defaultValue = "") String sex) {
-        userService.updateUser(id, username, age, sex);
-        return new ResponseEntity<>("User successfully updated", HttpStatus.OK);
+        if (userService.getUserById(id) != null) {
+            userService.updateUser(id, username, age, sex);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("{id}")
     public ResponseEntity<String> deleteUser(@PathVariable(value="id") long id) {
-        userService.deleteUser(id);
-        return new ResponseEntity<>("User successfully deleted", HttpStatus.OK);
+        if (userService.getUserById(id) == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else {
+            userService.deleteUser(id);
+            Optional<User> user = userService.getUserById(id);
+            return user == null
+                    ? new ResponseEntity<>(HttpStatus.OK)
+                    : new ResponseEntity<>(HttpStatus.SERVICE_UNAVAILABLE);
+        }
     }
 
     @GetMapping("{id}/info")
     public ResponseEntity<UserDto> getUserInfo(@PathVariable(value="id") long id) {
-        return ResponseEntity.ok(userService.getUserDto(id));
+        Optional<User> user = userService.getUserById(id);
+        return user != null
+                ? new ResponseEntity<>(userService.getUserDto(id), HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
